+++ tests/test_structure.py
@@ -189,5 +189,36 @@ def test_content_structure() -> None:
                 assert obj.parent.role in ("P", "H1", "H2", "H3")
 
 
+def test_element_hash():
+    with playa.open(TESTDIR / "pdf_structure.pdf") as pdf:
+        elements = set(pdf.structure.find_all())
+        page = pdf.pages[0]
+        for element in page.structure:
+            if element is not None:
+                if element.parent is not None:
+                    sibs = set(element.parent)
+                    assert element in sibs
+                    # Make sure that we can test for inequality too
+                    if len(sibs) > 1:
+                        assert sum(1 for sib in sibs if sib == element) < len(sibs)
+                assert element in elements
+                assert element == element
+
+
+def test_page_structure() -> None:
+    with playa.open(TESTDIR / "pdf_structure.pdf") as pdf:
+        assert pdf.structure is not None
+        elements = set(pdf.structure.find_all("Table"))
+        assert len(elements) == 1
+        page = pdf.pages[0]
+        count = 0
+        for element in page.structure.find_all("Table"):
+            assert element in elements
+            count += 1
+        assert count == 1
+        assert list(page.structure[0:5].find_all("Table")) == []
+        assert len(list(page.structure[-5:].find_all("Table"))) == 1
+
+
 if __name__ == "__main__":
     test_specific_structure()